+++ src/src/settings/settings_page.cpp
@@ -26,6 +26,7 @@
 #include <QLabel>
 #include <QLineEdit>
 #include <QMessageBox>
+#include <QPlainTextEdit>
 #include <QPushButton>
 #include <QScrollArea>
 #include <QSignalBlocker>
@@ -148,6 +149,7 @@ class settings_widget_t : public Core::IOptionsPageWidget
         int vector_search_max_indexing_threads = 0;
         bool debug_logging = false;
         bool detailed_request_logging = false;
+        QString system_prompt;
         bool inline_diff_refinement_enabled = false;
         bool agent_debug = false;
         qtmcp::server_definitions_t mcp_servers;
@@ -216,6 +218,7 @@ public:
         this->qdrantTestConnectionButton = this->ui->qdrantTestConnectionButton;
         this->debugLoggingCheck = this->ui->debugLoggingCheck;
         this->detailedRequestLoggingCheck = this->ui->detailedRequestLoggingCheck;
+        this->systemPromptEdit = this->ui->systemPromptEdit;
         this->inlineDiffRefinementCheck = this->ui->inlineDiffRefinementCheck;
         this->agentDebugCheck = this->ui->agentDebugCheck;
         this->mcpServersWidget = new mcp_servers_widget_t(this);
@@ -467,6 +470,7 @@ public:
 
         this->debugLoggingCheck->setChecked(s.debug_logging);
         this->detailedRequestLoggingCheck->setChecked(s.detailed_request_logging);
+        this->systemPromptEdit->setPlainText(s.system_prompt);
         this->inlineDiffRefinementCheck->setChecked(s.inline_diff_refinement_enabled);
 
         this->agentDebugCheck->setChecked(s.agent_debug);
@@ -527,6 +531,7 @@ public:
         connect(this->completionModelCombo, &QComboBox::currentTextChanged,
                 Utils::checkSettingsDirty);
         connect(this->detailedRequestLoggingCheck, &QCheckBox::toggled, Utils::checkSettingsDirty);
+        connect(this->systemPromptEdit, &QPlainTextEdit::textChanged, Utils::checkSettingsDirty);
         connect(this->inlineDiffRefinementCheck, &QCheckBox::toggled, Utils::checkSettingsDirty);
         connect(this->vectorSearchEnabledCheck, &QCheckBox::toggled, Utils::checkSettingsDirty);
         connect(this->mcpServersWidget, &mcp_servers_widget_t::changed, Utils::checkSettingsDirty);
@@ -543,6 +548,7 @@ public:
         s = this->settingsFromUi();
         s.debug_logging = this->debugLoggingCheck->isChecked();
         s.detailed_request_logging = this->detailedRequestLoggingCheck->isChecked();
+        s.system_prompt = this->systemPromptEdit->toPlainText().trimmed();
         s.inline_diff_refinement_enabled = this->inlineDiffRefinementCheck->isChecked();
         s.agent_debug = this->agentDebugCheck->isChecked();
         s.save();
@@ -627,6 +633,7 @@ private:
         s.vector_search_max_indexing_threads = this->vectorSearchMaxIndexingThreadsSpin->value();
         s.debug_logging = this->debugLoggingCheck->isChecked();
         s.detailed_request_logging = this->detailedRequestLoggingCheck->isChecked();
+        s.system_prompt = this->systemPromptEdit->toPlainText().trimmed();
         s.inline_diff_refinement_enabled = this->inlineDiffRefinementCheck->isChecked();
         s.agent_debug = this->agentDebugCheck->isChecked();
         s.mcp_servers = this->mcpServersWidget->servers();
@@ -765,6 +772,7 @@ private:
             this->vectorSearchMaxIndexingThreadsSpin->value(),
             this->debugLoggingCheck->isChecked(),
             this->detailedRequestLoggingCheck->isChecked(),
+            this->systemPromptEdit->toPlainText().trimmed(),
             this->inlineDiffRefinementCheck->isChecked(),
             this->agentDebugCheck->isChecked(),
             this->mcpServersWidget->servers(),
@@ -838,6 +846,7 @@ private:
 
         this->debugLoggingCheck->setChecked(snap.debug_logging);
         this->detailedRequestLoggingCheck->setChecked(snap.detailed_request_logging);
+        this->systemPromptEdit->setPlainText(snap.system_prompt);
         this->inlineDiffRefinementCheck->setChecked(snap.inline_diff_refinement_enabled);
         this->agentDebugCheck->setChecked(snap.agent_debug);
         this->mcpServersWidget->set_servers(snap.mcp_servers);
@@ -891,6 +900,7 @@ private:
     QPushButton *qdrantTestConnectionButton;
     QCheckBox *debugLoggingCheck;
     QCheckBox *detailedRequestLoggingCheck;
+    QPlainTextEdit *systemPromptEdit;
     QCheckBox *inlineDiffRefinementCheck;
     QCheckBox *agentDebugCheck;
     mcp_servers_widget_t *mcpServersWidget = nullptr;