+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 - Added an optional multi-turn inline diff refinement flow
+- Added system prompt
 
 ## [0.0.7] - 2026-03-21